requirements.txt
python-dotenv~=0.19.0
lxml~=4.6.2
aiohttp~=3.7.3
schedule~=1.1.0